Common code for rendering common .bed options. Add numbers per line
