+++ Client/src/components/reporters/BedGeneReporterForm.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { RadioList, CheckboxList, SingleSelect, TextBox, Checkbox } from '@veupathdb/wdk-client/lib/Components';
-import { deflineFieldOptions, GenomicSequenceRegionInputs, ProteinSequenceRegionInputs, FeaturesList, ComponentsList } from './BedFormElements';
+import { RadioList, CheckboxList, SingleSelect, TextBox, Checkbox, NumberSelector } from '@veupathdb/wdk-client/lib/Components';
+import { FeaturesList, ComponentsList } from './BedFormElements';
 import * as ComponentUtils from '@veupathdb/wdk-client/lib/Utils/ComponentUtils';
 import * as ReporterUtils from '@veupathdb/wdk-client/lib/Views/ReporterForm/reporterUtils';
 import SrtHelp from '../SrtHelp';
+import createBedForm from './BedFormFactory';
 
 let util = Object.assign({}, ComponentUtils, ReporterUtils);
-
 
 let splicedGenomicOptions = [
   { value: 'cds', display: 'CDS'},
@@ -27,16 +27,112 @@ let geneComponentOptions = [
   { value: 'intron', display: 'intron' },
 ];
 
+let genomicAnchorValues = [
+  { value: 'Start', display: 'Transcription Start***' },
+  { value: 'CodeStart', display: 'Translation Start (ATG)' },
+  { value: 'CodeEnd', display: 'Translation Stop Codon' },
+  { value: 'End', display: 'Transcription Stop***' }
+];
+
+let proteinAnchorValues = [
+  { value: 'Start', display: 'Downstream from Start' },
+  { value: 'End', display: 'Upstream from End' }
+];
+
+let signs = [
+  { value: 'plus', display: '+' },
+  { value: 'minus', display: '-' }
+];
+
+let SequenceRegionRange = props => {
+  let { label, anchor, sign, offset, formState, getUpdateHandler } = props;
+  return (
+    <React.Fragment>
+      <span>{label}</span>
+      <SingleSelect name={anchor} value={formState[anchor]}
+          onChange={getUpdateHandler(anchor)} items={genomicAnchorValues}/>
+      <SingleSelect name={sign} value={formState[sign]}
+          onChange={getUpdateHandler(sign)} items={signs}/>
+      <NumberSelector name={offset} value={formState[offset]}
+          start={0} end={10000} step={1}
+          onChange={getUpdateHandler(offset)} size="6"/>
+      nucleotides
+    </React.Fragment>
+  );
+};
+
+let ProteinRegionRange = props => {
+  let { label, anchor, offset, formState, getUpdateHandler } = props;
+  return (
+    <React.Fragment>
+      <span>{label}</span>
+      <SingleSelect name={anchor} value={formState[anchor]}
+          onChange={getUpdateHandler(anchor)} items={proteinAnchorValues}/>
+      <NumberSelector name={offset} value={formState[offset]}
+          start={0} end={10000} step={1}
+          onChange={getUpdateHandler(offset)} size="6"/>
+      amino acids
+    </React.Fragment>
+  );
+};
+
+let GenomicSequenceRegionInputs = props => {
+  let { formState, getUpdateHandler } = props;
+  return (
+    <div>
+      <div style={{marginLeft:"0.75em"}}>
+        <Checkbox name="reverseAndComplement" value={formState.reverseAndComplement} onChange={getUpdateHandler('reverseAndComplement')}/> Reverse & Complement
+      </div>
+      <div
+        style={{
+          display: 'inline-grid',
+          gridTemplateColumns: 'repeat(5, auto)',
+          alignItems: 'center',
+          gridRowGap: '0.25em',
+          gridColumnGap: '0.5em',
+          marginLeft: '0.75em'
+        }}
+      >
+        <SequenceRegionRange label="Begin at" anchor="upstreamAnchor" sign="upstreamSign"
+          offset="upstreamOffset" formState={formState} getUpdateHandler={getUpdateHandler}/>
+        <SequenceRegionRange label="End at" anchor="downstreamAnchor" sign="downstreamSign"
+          offset="downstreamOffset" formState={formState} getUpdateHandler={getUpdateHandler}/>
+      </div>
+    </div>
+  );
+}
+let ProteinSequenceRegionInputs = props => {
+  let { formState, getUpdateHandler } = props;
+  return (
+    <div>
+      <div
+        style={{
+          display: 'inline-grid',
+          gridTemplateColumns: 'repeat(4, auto)',
+          alignItems: 'center',
+          gridRowGap: '0.25em',
+          gridColumnGap: '0.5em',
+          marginLeft: '0.75em'
+        }}
+      >
+        <ProteinRegionRange label="Begin at" anchor="startAnchor3" offset="startOffset3"
+          formState={formState} getUpdateHandler={getUpdateHandler}/>
+        <ProteinRegionRange label="End at" anchor="endAnchor3" offset="endOffset3"
+          formState={formState} getUpdateHandler={getUpdateHandler}/>
+      </div>
+    </div>
+  );
+};
 
 /** @type import('./Types').ReporterFormComponent */
-let FastaGeneReporterForm = props => {
+let formBeforeCommonOptions = props => {
   let { formState, updateFormState, onSubmit, includeSubmit } = props;
   let getUpdateHandler = fieldName => util.getChangeHandler(fieldName, updateFormState, formState);
   let typeUpdateHandler = function(newTypeValue) {
     updateFormState(Object.assign({}, formState, { type: newTypeValue }));
   };
   return (
-    <div>
+    <React.Fragment>
       <h3>Choose the type of result:</h3>
       <div style={{marginLeft:"2em"}}>
         <RadioList name="type" value={formState.type}
@@ -58,31 +154,14 @@ let FastaGeneReporterForm = props => {
               <ComponentsList field="geneComponents" features={geneComponentOptions} formState={formState} getUpdateHandler={getUpdateHandler} />
             )},
           ]  
-          
           }/>
       </div>
-      <h3>Download Type:</h3>
-      <div style={{marginLeft:"2em"}}>
-        <RadioList name="attachmentType" value={formState.attachmentType}
-          onChange={getUpdateHandler('attachmentType')} items={util.attachmentTypes}/>
-      </div>
-      <h3>Fasta defline:</h3>
-      <div style={{marginLeft:"2em"}}>
-        <RadioList name="deflineType" value={formState.deflineType}
-          onChange={getUpdateHandler('deflineType')} items={[
-            {  value: "short", display: 'ID Only' },
-            { 
-              value: "full", display: 'Full Fasta Header',
-              body: (<ComponentsList field="deflineFields" features={deflineFieldOptions} formState={formState} getUpdateHandler={getUpdateHandler} />),
-            }
-            ]}/>
-      </div>
-      { includeSubmit &&
-        <div style={{margin:'0.8em'}}>
-          <button className="btn" type="submit" onClick={onSubmit}>Get Sequences</button>
-        </div>
-      }
-
+    </React.Fragment>
+  );
+};
+let formAfterSubmitButton = props => {
+  return (
+    <React.Fragment>
       <div>
         <hr/>
         <b>Note:</b><br/>
@@ -94,36 +173,30 @@ let FastaGeneReporterForm = props => {
         <hr/>
       </div>
       <SrtHelp/>
-    </div>
+    </React.Fragment>
   );
 };
+let getFormInitialState = () => ({
+  type: 'genomic',
 
-FastaGeneReporterForm.getInitialState = () => ({
-  formState: {
-    attachmentType: 'plain',
-    type: 'genomic',
-    reverseAndComplement: false,
-    deflineType: "short",
-    deflineFields: deflineFieldOptions.map((x) => x.value),
+  // sequence region inputs for 'genomic'
+  reverseAndComplement: false,
+  upstreamAnchor: 'Start',
+  upstreamSign: 'plus',
+  upstreamOffset: 0,
+  downstreamAnchor: 'End',
+  downstreamSign: 'plus',
+  downstreamOffset: 0,
 
-    // sequence region inputs for 'genomic'
-    upstreamAnchor: 'Start',
-    upstreamSign: 'plus',
-    upstreamOffset: 0,
-    downstreamAnchor: 'End',
-    downstreamSign: 'plus',
-    downstreamOffset: 0,
+  // sequence region inputs for 'protein'
+  startAnchor3: 'Start',
+  startOffset3: 0,
+  endAnchor3: 'End',
+  endOffset3: 0,
 
-    // sequence region inputs for 'protein'
-    startAnchor3: 'Start',
-    startOffset3: 0,
-    endAnchor3: 'End',
-    endOffset3: 0,
-    geneComponents: geneComponentOptions.map((x) => x.value),
-    proteinFeature: proteinFeatureOptions[0].value,
-    splicedGenomic: splicedGenomicOptions[0].value,
-  },
-  formUiState: {}
+  geneComponents: geneComponentOptions.map((x) => x.value),
+  proteinFeature: proteinFeatureOptions[0].value,
+  splicedGenomic: splicedGenomicOptions[0].value,
 });
 
-export default FastaGeneReporterForm;
+export default createBedForm(formBeforeCommonOptions, formAfterSubmitButton, getFormInitialState);
